Backend/pyproject.toml
[tool.poetry]
name = "9erp-flask"
version = "0.1.0"
description = "Backend flask application for 9 education."
authors = ["teja <tejabhargavpodila@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
connexion = {extras = ["swagger-ui"], version = "^2.14.2"}
python-dateutil = "^2.8.2"
swagger-ui-bundle = "^0.0.9"
flask-pymongo = "2.3.0"
python-dotenv = ">=0.19.2,<=0.21.1"
loguru = "^0.7.2"
flask-cors = "^4.0.0"
pandas = "^2.1.4"
openpyxl = "^3.1.2"
flask-mail = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
